requirements.txt
Django==1.4
PIL==1.1.7
South==0.7.5
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.1
django-avatar==1.0.5
django-guardian==1.0.4
django-pagination==1.0.7
django-tagging==0.3.1
django-threadedcomments==0.5.3
django-uni-form==0.9.0
django-userena==1.1.2
easy-thumbnails==1.0.3
psycopg2==2.4.5
wsgiref==0.1.2

--extra-index-url=http://dist.pinaxproject.com/dev/
django-groups